Crate burn_train

A library for training neural networks using the burn crate.

Modules

checkpoint

The checkpoint module.

logger

The logger module.

metric

The metric module.

train

The trainer module.

Structs

AsyncTrainerCallback

Async trainer callback tracker.

ClassificationOutput

Simple classification output adapted for multiple metrics.

Learner

Learner struct encapsulating all components necessary to train a Neural Network model.

LearnerBuilder

Struct to configure and create a learner.

LearnerItem

A learner item.

MultiDevicesTrainStep

Multi devices train step.

RegressionOutput

Simple regression output adapted for multiple metrics.

TrainEpoch

A training epoch.

TrainOutput

A training output.

ValidEpoch

A validation epoch.

Traits

LearnerCallback

The base trait for trainer callbacks.

TrainStep

Trait to be implemented for training models.

ValidStep

Trait to be implemented for validating models.