Struct burn_train::LearnerBuilder

pub struct LearnerBuilder<B, T, V, M, O, S>where
    T: Send + Sync + 'static,
    V: Send + Sync + 'static,
    B: ADBackend,
    M: ADModule<B>,
    O: Optimizer<M, B>,
    S: LRScheduler,{ /* private fields */ }

Struct to configure and create a learner.

Implementations

impl<B, T, V, Model, Optim, LR> LearnerBuilder<B, T, V, Model, Optim, LR>where T: Send + Sync + 'static, V: Send + Sync + 'static, B: ADBackend, Model: ADModule<B>, Optim: Optimizer<Model, B>, LR: LRScheduler,

pub fn new(directory: &str) -> Self

Creates a new learner builder.

Arguments

directory - The directory to save the checkpoints.

pub fn metric_loggers<MT, MV>(self, logger_train: MT, logger_valid: MV) -> Selfwhere MT: MetricLogger + 'static, MV: MetricLogger + 'static,

Replace the default metric loggers with the provided ones.

Arguments

logger_train - The training logger.

logger_valid - The validation logger.

pub fn renderer<DR>(self, renderer: DR) -> Selfwhere DR: DashboardRenderer + 'static,

Replace the default CLI renderer with a custom one.

Arguments

renderer - The custom renderer.

pub fn metric_train<M: Metric + 'static>(self, metric: M) -> Selfwhere T: Adaptor<M::Input>,

Register a training metric.

pub fn metric_valid<M: Metric + 'static>(self, metric: M) -> Selfwhere V: Adaptor<M::Input>,

Register a validation metric.

pub fn grads_accumulation(self, accumulation: usize) -> Self

Enable gradients accumulation.

Notes

When you enable gradients accumulation, the gradients object used by the optimizer will be the sum of all gradients generated by each backward pass. It might be a good idea to reduce the learning to compensate.

The effect is similar to increasing the batch size and the learning rate by the accumulation amount.

pub fn metric_train_plot<M>(self, metric: M) -> Selfwhere M: Metric + Numeric + 'static, T: Adaptor<M::Input>,

Register a training metric and displays it on a plot.

Notes

Only numeric metric can be displayed on a plot. If the same metric is also registered for the validation split, the same graph will be used for both.

pub fn metric_valid_plot<M: Metric + Numeric + 'static>(self, metric: M) -> Selfwhere V: Adaptor<M::Input>,

Register a validation metric and displays it on a plot.

Notes

Only numeric metric can be displayed on a plot. If the same metric is also registered for the training split, the same graph will be used for both.

pub fn num_epochs(self, num_epochs: usize) -> Self

The number of epochs the training should last.

pub fn devices(self, devices: Vec<B::Device>) -> Self

Run the training loop on multiple devices.

pub fn checkpoint(self, checkpoint: usize) -> Self

The epoch from which the training must resume.

pub fn log_to_file(self, enabled: bool) -> Self

By default, Rust logs are captured and written into experiment.log. If disabled, standard Rust log handling will apply.

pub fn with_file_checkpointer<FR>(self, num_keep: usize, recorder: FR) -> Selfwhere FR: FileRecorder + 'static,

Register a checkpointer that will save the optimizer and the model.

The number of checkpoints to be keep should be set to a minimum of two to be safe, since they are saved and deleted asynchronously and a crash during training might make a checkpoint non-usable.

pub fn build( self, model: Model, optim: Optim, lr_scheduler: LR ) -> Learner<B, Model, Optim, LR, T, V>where Model::Record: 'static, Optim::Record: 'static, LR::Record: 'static,

Create the learner from a model and an optimizer. The learning rate scheduler can also be a simple learning rate.