Crate burn_train

A library for training neural networks using the burn crate.

Modules

checkpoint

The checkpoint module.

logger

The logger module.

metric

The metric module.

renderer

Renderer modules to display metrics and training information.

train

The trainer module.

Structs

ClassificationOutput

Simple classification output adapted for multiple metrics.

Learner

Learner struct encapsulating all components necessary to train a Neural Network model.

LearnerBuilder

Struct to configure and create a learner.

MetricEarlyStoppingStrategy

An early stopping strategy based on a metrics collected during training or validation.

MultiDevicesTrainStep

Multi devices train step.

RegressionOutput

Simple regression output adapted for multiple metrics.

TrainEpoch

A training epoch.

TrainOutput

A training output.

TrainingInterrupter

A handle that allows aborting the training process early.

ValidEpoch

A validation epoch.

Enums

StoppingCondition

The condition that early stopping strategies should follow.

Traits

EarlyStoppingStrategy

A strategy that checks if the training should be stopped.

TrainStep

Trait to be implemented for training models.

ValidStep

Trait to be implemented for validating models.